Change internal repr
We need to do this in order to faithfully represent the analysis
of multiple characters.

For example, if the answer is 'usage', then the correct response
should be: [ad]AGE, indicating that 'a' is impossible in the first
position, but correct in the second.

Equally, if the correct word is "foody", and the guess is "foamo",
then the correct response is "FO[am]o", indicating that the second
'o' is misplaced.

In order to get this right, we change the representation of the
knowledge from a set of facts to a state-of-the-world, and we
process this statefully when printing guesses, consuming misplaced
tokens using a fold.

We track three pieces of data:

- the letters we have placed correctly
- the number of each letter that the word must have at least (including
  correctly guessed letters, and misplaced ones)
- places where we know letters cannot go

From a combination of the at-least counts, and the exclusion lists, we
have enough information to answer questions about placement of misplaced
letters.